Intent 연결

+++ app/src/main/java/com/example/dto/DeviceDetailDTO.java
@@ -7,7 +7,7 @@ import com.example.com02.R;
 public class DeviceDetailDTO {
     private String deviceName;
     private String category;    //분류
-    private Double mAh;  //배터리 용량
+    private Integer mAh;  //배터리 용량
     private Double maximum_output;    //최대 출력 와트
     private String contents;    //충전 단자
     private String usingTime;     //전체 사용 가능 시간 (원래 type: Time)
@@ -15,7 +15,7 @@ public class DeviceDetailDTO {
     private String averageDays;
     private Long capacityOfUserId;
 
-    public DeviceDetailDTO(String deviceName, String category, Double mAh, Double maximum_output,
+    public DeviceDetailDTO(String deviceName, String category, Integer mAh, Double maximum_output,
                            String contents, String usingTime, String nowCapacity, String averageDays,Long capacityOfUserId) {
         this.deviceName = deviceName;
         this.category = category;
@@ -34,7 +34,7 @@ public class DeviceDetailDTO {
     public String getCategory() {
         return category;
     }
-    public Double getmAh(){return mAh;}
+    public Integer getmAh(){return mAh;}
     public Double getMaximum_output() {
         return maximum_output;
     }
